requirements.txt
Django==2.0.4
numpy==1.14.2
Pillow==5.1.0
pytz==2018.3
scikit-learn==0.19.1
scipy==1.0.1
pydot==1.3.0
